// web/gen/api.ts
export interface PredictionIntentResponse {
    /**
     * @generated from protobuf field: string tx_id = 1
     */
    txId: string;
    /**
     * @generated from protobuf field: string net = 2
     */
    net: string;
    /**
     * @generated from protobuf field: string market_id = 3
     */
    marketId: string;
    /**
     * @generated from protobuf field: string generated_at = 4
     */
    generatedAt: string; // when the API accepted the order
    /**
     * @generated from protobuf field: string account_id = 5
     */
    accountId: string;
    /**
     * @generated from protobuf field: uint32 side = 9
     */
    side: number;
    /**
     * @generated from protobuf field: uint32 action = 10
     */
    action: number;
    /**
     * @generated from protobuf field: uint64 limit_yes_price = 11
     */
    limitYesPrice: bigint;
    /**
     * @generated from protobuf field: uint64 qty_shares = 12
     */
    qtyShares: bigint;
    /**
     * @generated from protobuf field: uint64 collateral_cap = 13
     */
    collateralCap: bigint;
    /**
     * @generated from protobuf field: uint64 deadline = 14
     */
    deadline: bigint;
    /**
     * @generated from protobuf field: uint64 shares_filled = 15
     */
    sharesFilled: bigint;
    /**
     * @generated from protobuf field: uint64 collateral_filled = 16
     */
    collateralFilled: bigint;
}
// ...
export const PredictionIntentResponse = new PredictionIntentResponse$Type();